feat(playground): setup playground system
All the pages inside the `pages` directory can be used as a playground.
We have a custom `_app.tsx` file which ensures that there is a "shell"
for each component.

If you go to a route that is essentially a folder, you will get a page
with a listing of all the examples available. We can do this by parsing
the file system tree and display a nested tree on the page for this
(sub) section.